% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRtoSignac.R
\name{addCustomizeDimRed}
\alias{addCustomizeDimRed}
\title{addCustomizeDimRed}
\usage{
addCustomizeDimRed(
  ArchRProject,
  SeuratObject,
  addUMAPs = "UMAP",
  reducedDims = "IterativeLSI",
  reducedDimsType = "IterativeLSI"
)
}
\arguments{
\item{ArchRProject}{An ArchRProject}

\item{SeuratObject}{A Seurat object}

\item{addUMAPs}{add UMAPs of the selected UMAPs}

\item{reducedDims}{the selected dimension reduction to be transfered from ArchRProject to Signac SeuratObject}

\item{reducedDimsType}{'Harmony' or 'IterativeLSI', this parameter tells the type of the reduced dimension (default is 'IterativeLSI')}
}
\description{
This function adds dimension reduction ('Harmony' and/or 'IterativeLSI') and UMAP to a SeuratObject with customized name of the dimension
}
\examples{
seurat_atac <- addCustomizeDimRed(ArchRProject = proj, SeuratObject = seurat_atac, addUMAPs = "UMAP", reducedDims = 'IterativeLSI', reducedDimsType = 'IterativeLSI') #reducedDims == c('IterativeLSI', 'Harmony')
}
