% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRtoSignac.R
\name{addDimRed}
\alias{addDimRed}
\title{addDimRed}
\usage{
addDimRed(
  ArchRProject = NULL,
  SeuratObject = NULL,
  reducedDims = "IterativeLSI"
)
}
\arguments{
\item{ArchRProject}{A ArchRProject}

\item{SeuratObject}{A Seurat object}

\item{reducedDims}{A reduction dimension can be transfered from from ArchRProject to Signac SeuratObject}
}
\description{
This function adds dimension reduction, 'Harmony' or 'IterativeLSI' and UMAP to SeuratObject
}
\examples{
seurat_atac <- addDimRed(ArchRProject = proj, SeuratObject = seurat_atac,reducedDims = 'IterativeLSI')
}
