% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRtoSignac.R
\name{addTwoDimRed}
\alias{addTwoDimRed}
\title{addTwoDimRed}
\usage{
addTwoDimRed(
  ArchRProject,
  SeuratObject,
  addUMAPs = "UMAP",
  reducedDims1 = NULL,
  reducedDims2 = NULL
)
}
\arguments{
\item{ArchRProject}{An ArchRProject}

\item{SeuratObject}{A Seurat object}

\item{addUMAPs}{add UMAPs by default}

\item{reducedDims1}{'Harmony'or 'IterativeLSI' or 'IterativeLSI2', the dimension reduction to be transfered from ArchRProject to Signac SeuratObject (default is 'IterativeLSI')}

\item{reducedDims2}{'Harmony'or 'IterativeLSI' or 'IterativeLSI2', the dimension reduction to be transfered from ArchRProject to Signac SeuratObject (default is 'IterativeLSI')}
}
\description{
This function adds dimension reduction ('Harmony' and/or 'IterativeLSI') and UMAP to a SeuratObject
}
\examples{
seurat_atac <- addTwoDimRed(ArchRProject = proj, SeuratObject = seurat_atac, reducedDims1 = 'IterativeLSI', reducedDims2= 'Harmony')
}
