% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRtoSignac.R
\name{ArchR2Signac}
\alias{ArchR2Signac}
\title{ArchR2Signac}
\usage{
ArchR2Signac(
  ArchRProject,
  refversion,
  samples = NULL,
  fragments_dir = NULL,
  pm,
  output_dir = "/outs/",
  annotation
)
}
\arguments{
\item{ArchRProject}{An ArchRProject}

\item{refversion}{The assembly release and versions of UCSC genome reference}

\item{samples}{List of all the samples from the ArchRProject}

\item{fragments_dir}{PATH to the cellranger-atac output--the folder that contains all samples folders, not the one with '/outs/fragments.tsv.gz'.}

\item{pm}{Peak matrix (output) from the function getPeakMatrix}

\item{output_dir}{PATH before 'fragments.tsv.gz' from cell-ranger-atac (typically this is '/outs/' and is the default)}

\item{annotation}{annotation from the function getAnnotation()}
}
\description{
This function converts the ArchRProject to a SeuratObject by creating a list of Seurat objects for each sample with their corresponding peak matrix and then merging all of the objects in the list.
}
\examples{
seurat_atac <- ArchR2Signac(ArchRProject = proj1, refversion = 'hg38', samples = samples, fragments_dir = fragments_dir, pm = pm, output_dir = '/outs/', annotation = annotations)
}
