% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRtoSignac.R
\name{getGeneScoreMatrix}
\alias{getGeneScoreMatrix}
\title{getGeneScoreMatrix}
\usage{
getGeneScoreMatrix(ArchRProject = NULL, SeuratObject = NULL)
}
\arguments{
\item{ArchRProject}{A ArchRProject}

\item{SeuratObject}{A Seurat object}
}
\description{
This function gets gene score matrix from ArchR project and change the row names of gene score matrix to their matched gene features
}
\examples{
gsm <- getGeneScoreMatrix(ArchRProject = proj, SeuratObject = seurat_atac)
}
