% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRtoSignac.R
\name{getPeakMatrix}
\alias{getPeakMatrix}
\title{getPeakMatrix}
\usage{
getPeakMatrix(ArchRProject)
}
\arguments{
\item{ArchRProject}{An ArchRProject}
}
\description{
This function gets the fixed-width peak matrix from the ArchRProject and changes the row names of the peak matrix to their matched chromosome range.
}
\examples{
pm <- getPeakMatrix(proj)
}
