% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRtoSignac.R
\name{getpm}
\alias{getpm}
\title{getpm}
\usage{
getpm(ArchRProject = NULL)
}
\arguments{
\item{ArchRProject}{A ArchRProject}
}
\description{
This function gets fixed-width peak matrix from ArchR project and change the row names of peak matrix to their matched chromosome range
}
\examples{
pm <- getpm(proj)
}
