% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRtoSignac.R
\name{loadinglibrary}
\alias{loadinglibrary}
\title{loadinglibrary}
\usage{
loadinglibrary(packages)
}
\arguments{
\item{packages}{Vector of libraries required for ArchRtoSignac (most dependencies should be installed when installing ArchRtoSignac)}
}
\description{
This function is a quick way to check if the required packages for ArchRtoSignac package are installed. If the packages are available, they are loaded automatically.
}
\examples{
loadinglibrary(dependencies)
}
