% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRtoSignac.R
\name{addDimRed}
\alias{addDimRed}
\title{addDimRed}
\usage{
addDimRed(ArchRProject, SeuratObject, reducedDims = "IterativeLSI")
}
\arguments{
\item{ArchRProject}{An ArchRProject}

\item{SeuratObject}{A Seurat object}

\item{reducedDims}{'Harmony' or 'IterativeLSI', the dimension reduction to be transfered from ArchRProject to Signac SeuratObject (default is 'IterativeLSI')}
}
\description{
This function adds a dimension reduction (either 'Harmony' or 'IterativeLSI') and UMAP to a SeuratObject
}
\examples{
seurat_atac <- addDimRed(ArchRProject = proj, SeuratObject = seurat_atac, reducedDims = 'IterativeLSI')
}
