% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRtoSignac.R
\name{ArchR2Signac}
\alias{ArchR2Signac}
\title{ArchR2Signac}
\usage{
ArchR2Signac(
  ArchRProject,
  refversion,
  samples = NULL,
  fragments_dir = NULL,
  pm,
  fragments_fromcellranger = NULL,
  fragments_file_extension = NULL,
  annotation
)
}
\arguments{
\item{ArchRProject}{An ArchRProject}

\item{refversion}{The assembly release and versions of UCSC genome reference}

\item{samples}{List of all the samples from the ArchRProject}

\item{fragments_dir}{PATH to the cellranger-atac output--the folder that contains all samples folders, not the one with '/outs/fragments.tsv.gz'.}

\item{pm}{Peak matrix (output) from the function getPeakMatrix}

\item{fragments_fromcellranger}{This is an Yes or No selection ("NO" | "N" | "No" or "YES" | "Y" | "Yes")}

\item{fragments_file_extension}{File_Extension for fragments files (typically they should be '.tsv.gz' or '.fragments.tsv.gz')}

\item{annotation}{annotation from the function getAnnotation()}
}
\description{
This function converts the ArchRProject to a SeuratObject by creating a list of Seurat objects for each sample with their corresponding peak matrix and then merging all of the objects in the list.
}
\examples{
seurat_atac <- ArchR2Signac(ArchRProject = proj1, refversion = 'hg38', samples = samples, fragments_dir = fragments_dir, pm = pm, fragments_fromcellranger = "Yes", fragments_file_extension = NULL, annotation = annotations)
}
