% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRtoSignac.R
\name{getAnnotation}
\alias{getAnnotation}
\title{getAnnotation}
\usage{
getAnnotation(reference, seqStyle = "UCSC", refversion)
}
\arguments{
\item{reference}{An Ensembl genome reference used for the Signac function GetGRangesFromEnsDb to extract gene annotations from EnsDb (for example: EnsDb.Hsapiens.v86)}

\item{seqStyle}{Sequence style to change the annotation extracted from EnsDb to (default is ‘UCSC’)}

\item{refversion}{The assembly release and versions of UCSC genome reference (for example: 'hg38')}
}
\description{
This function gets the gene annotation, which includes information related to genomic locations and their associated annotations, from Ensembl Database in GRanges Object for the Seurat object.
Then it changes the annotation to the UCSC style (default).
}
\examples{
annotations <- getAnnotation(reference = EnsDb.Hsapiens.v86, seqStyle = 'UCSC', refversion = 'hg38')
}
