% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRtoSignac.R
\name{getGeneScoreMatrix}
\alias{getGeneScoreMatrix}
\title{getGeneScoreMatrix}
\usage{
getGeneScoreMatrix(ArchRProject, SeuratObject)
}
\arguments{
\item{ArchRProject}{An ArchRProject}

\item{SeuratObject}{A Seurat object}
}
\description{
This function gets the gene score matrix from an ArchRProject and changes the row names of gene score matrix to their matched gene features
}
\examples{
gsm <- getGeneScoreMatrix(ArchRProject = proj, SeuratObject = seurat_atac)
}
