% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRtoSignac.R
\name{getPeakMatrix}
\alias{getPeakMatrix}
\title{getPeakMatrix}
\usage{
getPeakMatrix(ArchRProject = NULL)
}
\arguments{
\item{ArchRProject}{A ArchRProject}
}
\description{
This function gets fixed-width peak matrix from ArchR project and change the row names of peak matrix to their matched chromosome range
}
\examples{
pm <- getPeakMatrix(proj)
}
